backend/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
pytesseract
pillow
pydantic
python-multipart
pandas
requests
# YOLO dependencies
torch
torchvision
ultralytics
# RAG dependencies
sentence-transformers
openai
# Optional: ONNX runtime for ONNX models
# onnxruntime
# Optional dependencies, uncomment to use
# easyocr
# PyMuPDF
